requirements.txt
Django>=4.2
stripe>=5.0
python-dotenv>=1.0
dj-database-url>=1.0
psycopg2-binary>=2.9
